Rename production configuration
